baekjoon 1008 modified

+++ 백준/1008.kt
@@ -5,6 +5,6 @@ fun main(args: Array<String>){
     val br = BufferedReader(InputStreamReader(System.`in`))
     val token = StringTokenizer(br.readLine())
 
-    println((Integer.parseInt(token.nextToken()).toDouble()/  Integer.parseInt(token.nextToken())).toFloat().toDouble())
+    println(Integer.parseInt(token.nextToken()).toDouble()/  Integer.parseInt(token.nextToken()).toDouble())
     br.close() }
